+++ src/main.js
@@ -1,0 +1,3 @@
+import vcModal from './Modal'
+
+module.exports = vcModal